client/requirements.txt
# Client-side dependencies (installed on your laptop)
requests
pillow
